change var function in the same file
